+++ tests/App/AppKernel.php
@@ -36,24 +36,32 @@ class AppKernel extends Kernel
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         $loader->load(__DIR__.'/config.yml');
+
+        if (Kernel::MAJOR_VERSION >= 5) {
+            $loader->load(__DIR__.'/security_5.yml');
+            $loader->load(__DIR__.'/session_5.yml');
+        } else {
+            $loader->load(__DIR__.'/security_4.yml');
+            $loader->load(__DIR__.'/session_4.yml');
+        }
     }
 
-    public function getCacheDir()
+    public function getCacheDir(): string
     {
         return $this->getBaseDir().'cache';
     }
 
-    public function getLogDir()
+    public function getLogDir(): string
     {
         return $this->getBaseDir().'log';
     }
 
-    protected function getBaseDir()
+    protected function getBaseDir(): string
     {
         return sys_get_temp_dir().'/LiipFunctionalTestBundle/'.(new \ReflectionClass($this))->getShortName().'/var/';
     }
 
-    public function getProjectDir()
+    public function getProjectDir(): string
     {
         return __DIR__;
     }