+++ tests/App/Controller/DefaultController.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Liip\Acme\Tests\App\Controller;
 
+use Doctrine\ORM\EntityManagerInterface;
 use Liip\Acme\Tests\App\Entity\User;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -30,10 +31,10 @@ class DefaultController extends AbstractController
         );
     }
 
-    public function userAction(int $userId): Response
+    public function userAction(EntityManagerInterface $entityManager, int $userId): Response
     {
         /** @var \Liip\Acme\Tests\App\Entity\User $user */
-        $user = $this->getDoctrine()
+        $user = $entityManager
             ->getRepository(User::class)
             ->find($userId);
 
@@ -74,16 +75,17 @@ class DefaultController extends AbstractController
 
         $form->handleRequest($request);
 
+        $flashMessage = null;
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->get('session')->getFlashBag()->add(
-                'notice',
-                'Name submitted.'
-            );
+            $flashMessage = 'Name submitted.';
         }
 
         return $this->render(
             $template,
-            ['form' => $form->createView()]
+            [
+                'form' => $form->createView(),
+                'flash_message' => $flashMessage,
+            ]
         );
     }